# api/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/topic/v1/topic.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/addons/identify": {
      "post": {
        "summary": "openapi addons identify",
        "operationId": "Identify",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1AddonsIdentifyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AddonsIdentifyRequest"
            }
          }
        ],
        "tags": [
          "Openapi"
        ]
      }
    },
    "/identify": {
      "get": {
        "summary": "openapi identify",
        "operationId": "Identify",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1IdentifyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Openapi"
        ]
      }
    },
    "/status": {
      "get": {
        "summary": "openapi status",
        "operationId": "Status",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1StatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Openapi"
        ]
      }
    },
...
    "/tenant/disable": {
      "post": {
        "summary": "openapi tenant disable",
        "operationId": "TenantDisable",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1TenantDisableResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1TenantDisableRequest"
            }
          }
        ],
        "tags": [
          "Openapi"
        ]
      }
    },
    "/tenant/enable": {
      "post": {
        "summary": "openapi tenant enable",
        "operationId": "TenantEnable",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1TenantEnableResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1TenantEnableRequest"
            }
          }
        ],
        "tags": [
          "Openapi"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AddonsIdentifyRequest": {
      "type": "object",
      "properties": {
        "plugin": {
          "$ref": "#/definitions/v1BriefPluginInfo"
        },
        "implemented_addons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ImplementedAddons"
          }
        }
      },
      "description": "*\nThe request value of the POST /v1/addons/identify,\nused when the plugin in the registration process implements\nthe extension point of the registered plugin."
    },
    "v1AddonsIdentifyResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        }
      },
      "description": "*\nThe response value of the POST /v1/addons/identify,\nused when the plugin in the registration process implements\nthe extension point of the registered plugin."
    },
    "v1AddonsPoint": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        }
      },
      "description": "*\nmessage plugin declares extension point."
    },
    "v1BriefPluginInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      },
      "description": "*\nmessage brief plugin information."
    },
    "v1ConsoleEntry": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "icon": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "entry": {
          "type": "string"
        },
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ConsoleEntry"
          }
        }
      },
      "description": "plugin console entry."
    },
...
    "v1IdentifyResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        },
        "plugin_id": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "tkeel_version": {
          "type": "string"
        },
        "addons_point": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1AddonsPoint"
          }
        },
        "implemented_plugin": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ImplementedPlugin"
          }
        },
        "entries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ConsoleEntry"
          }
        }
      },
      "description": "*\nThe response value of the GET /v1/identify,\nused when the plugin registered."
    },
    "v1ImplementedAddons": {
      "type": "object",
      "properties": {
        "addons_point": {
          "type": "string"
        },
        "implemented_endpoint": {
          "type": "string"
        }
      },
      "description": "*\nmessage plugin declares implemented other plugin extension point."
    },
    "v1ImplementedPlugin": {
      "type": "object",
      "properties": {
        "plugin": {
          "$ref": "#/definitions/v1BriefPluginInfo"
        },
        "addons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ImplementedAddons"
          }
        }
      },
      "description": "*\nmessage plugin implemented other plugin."
    },
...
    "v1PluginStatus": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "WAIT_RUNNING",
        "ERR_REGISTER",
        "RUNNING"
      ],
      "default": "UNKNOWN",
      "description": "*\nplugin state enum."
    },
    "v1Result": {
      "type": "object",
      "properties": {
        "ret": {
          "$ref": "#/definitions/v1Retcode"
        },
        "msg": {
          "type": "string"
        }
      },
      "description": "*\nmessage result."
    },
    "v1Retcode": {
      "type": "string",
      "enum": [
        "OK",
        "BAD_REQEUST",
        "INTERNAL_ERROR"
      ],
      "default": "OK",
      "description": "*\nerror code enum."
    },
    "v1StatusResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        },
        "status": {
          "$ref": "#/definitions/v1PluginStatus"
        }
      },
      "description": "*\nThe request value of the GET /v1/Status.\nused when the tkeel checks the status of plugins regularly."
    },
...
    "v1TenantDisableRequest": {
      "type": "object",
      "properties": {
        "tenant_id": {
          "type": "string"
        },
        "extra": {
          "type": "string",
          "format": "byte"
        }
      },
      "description": "*\nThe request value of the POST /v1/tenant/disable,\nused when the tenant administrator disables the plugin."
    },
    "v1TenantDisableResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        }
      },
      "description": "*\nThe response value of the POST /v1/tenant/disable,\nused when the tenant administrator disables the plugin."
    },
    "v1TenantEnableRequest": {
      "type": "object",
      "properties": {
        "tenant_id": {
          "type": "string"
        },
        "extra": {
          "type": "string",
          "format": "byte"
        }
      },
      "description": "*\nThe request value of the POST /v1/tenant/enable,\nused when the tenant administrator enables the plugin."
    },
    "v1TenantEnableResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        }
      },
      "description": "*\nThe response value of the POST /v1/tenant/enable,\nused when the tenant administrator enables the plugin."
    },
...
  }
}